refactor: Replace withSearch() with SearchTerms component
