+++ translate/src/modules/search/components/SearchTerms.test.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { mount } from 'enzyme';
+
+import { SearchTerms } from './SearchTerms';
+
+describe('SearchTerms', () => {
+  it('does not break on regexp special characters', () => {
+    const wrapper = mount(
+      <SearchTerms search='(bar'>{'foo (bar)'}</SearchTerms>,
+    );
+    expect(wrapper.find('mark').text()).toEqual('(bar');
+  });
+});
